fix(scan): stop file-scan progress messages being dropped mid-scan
Reported: navigating into an issue's detail view mid file-scan (e.g.
at 4/12 pages) and back left the scan permanently stuck - "Scanning
page 12 of 12..." forever, Cancel doing nothing.

Root cause: SCAN_FILE_PAGE_ISSUES/SCAN_FILE_COMPLETE (and LOAD_ISSUES)
were only handled inside IssuesOverviewList's own message listener,
which unmounts whenever the user navigates to a different screen
(ISSUE_LIST_VIEW/TOUCH_TARGET_ISSUE_LIST_VIEW). A file scan keeps
running in the background regardless of which screen is showing, and
postMessage doesn't replay missed messages - so any pages that
finished while the user was elsewhere, including the final
SCAN_FILE_COMPLETE, were silently lost. `scanning` never flipped back
to false, and Cancel only ever toggled fileScanCancelled (already
true), so it looked like it did nothing.

Moved this handling to App's already-persistent top-level listener
(same place SCAN_FILE_PROGRESS already lived), which never unmounts
regardless of route. IssuesOverviewList no longer needs its own
listener at all.

Also fixed, found via this change: AltTextGenerator's own message
listener did `event.data.pluginMessage` with no null guard on
event.data itself - throws inside an async handler, surfacing as an
unhandled rejection that fails the whole vitest run rather than a
single test.
